+++ examples/complex.s2.php
@@ -33,8 +33,8 @@ print(sprintf('количество карт: %s' . PHP_EOL, $cards->count()));
 
 foreach ($cards as $card) {
     print('---------------' . PHP_EOL);
-    print(sprintf('получаем транзакции по карте с id %s' . PHP_EOL, $card->getCardId()->getId()));
-    // получение транзакций по карте
+    print(sprintf('получаем транзакции баллов по карте с id %s' . PHP_EOL, $card->getCardId()->getId()));
+    // получение транзакций бонусных баллов по карте
     $transactionCollection = $orgTransactionsTransport->getTransactionsByCard($card);
     foreach ($transactionCollection as $trx) {
         print(sprintf('-- %s | %s %s | %s' . PHP_EOL,
@@ -43,5 +43,26 @@ foreach ($cards as $card) {
             $trx->getSum()->getCurrency()->getCode(),
             $trx->getType()->getCode())
         );
+    }
+    // получение истории покупок по карте
+    $salesHistory = $orgTransactionsTransport->getSalesHistoryByCard($card);
+    print(sprintf('история покупок: ' . PHP_EOL));
+    foreach ($salesHistory as $historyItem) {
+        print(sprintf('%s | %s | %s %s' . PHP_EOL,
+            $historyItem->getLineNumber(),
+            $historyItem->getDocumentId()->getId(),
+            $historyItem->getSum()->getAmount(),
+            $historyItem->getSum()->getCurrency()->getCode()
+        ));
+        print('====== табличная часть: ' . PHP_EOL);
+        foreach ($historyItem->getProducts() as $productRow) {
+            print(sprintf('            %s | %s |  %s cnt | %s %s ' . PHP_EOL,
+                $productRow->getArticleId()->getId(),
+                $productRow->getName(),
+                $productRow->getQuantity(),
+                $productRow->getPrice()->getAmount(),
+                $productRow->getPrice()->getCurrency()
+            ));
+        }
     }
 }